backend/requirements.txt
Flask
flask-cors
werkzeug
langchain
pymilvus
langchain-community
langchain-unstructured
google-generative-ai
google-generativeai
requests
python-multipart
